requirements.txt
Pillow
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
transformers
timm
httpx
pyaudio
scipy
snac
opencv-python
accelerate